import_brevet/requirements.txt
--extra-index-url https://us-central1-python.pkg.dev/baltic-star-cloud/brevet-top-python-repo/simple/
firebase-admin==5.2.0
google-cloud-firestore==2.3.4
google-cloud-logging==2.6.0
Flask==2.2.5
flask-cors==3.0.10
python_dateutil==2.8.2
brevet-top-gcp-utils==0.1.4
brevet-top-misc-utils==0.1.0
brevet-top-plot-a-route==0.1.9
